1служба: 5я функция падает если что то в 1м состоянии

+++ services/1/schedule/schedule.h
@@ -20,7 +20,7 @@ class schedule
     job*  getJobByIndex(DWORD index);
     WORD getJobsQuantity();
 
-    errType cursorPos(WORD **ids);
+    errType cursorPos(BYTE **ids);
     errType removeAllJobsBefore(DWORD );
 
     errType run();
